chore: Move list folder to up-level directory
