+++ packages/artalk/src/list/index.ts
@@ -1,0 +1,3 @@
+import List from './list'
+
+export default List
