+++ packages/artalk/src/list/list-lite.ts
@@ -5,8 +5,8 @@ import * as Utils from '../lib/utils'
 import * as Ui from '../lib/ui'
 import Api from '../api'
 import Comment from '../comment'
-import Pagination from './pagination'
-import ReadMoreBtn from './read-more-btn'
+import Pagination from '../components/pagination'
+import ReadMoreBtn from '../components/read-more-btn'
 import * as ListNest from './list-nest'
 import { backendMinVersion } from '../../package.json'
