+++ packages/artalk/src/list/list.ts
@@ -5,7 +5,7 @@ import Context from '~/types/context'
 import * as Utils from '../lib/utils'
 import * as Ui from '../lib/ui'
 import Api from '../api'
-import ListHTML from './html/list.html?raw'
+import ListHTML from './list.html?raw'
 import ListLite from './list-lite'
 
 export default class List extends ListLite {
